+++ Pinta.Resources/Icons.cs
@@ -42,7 +42,6 @@ public static class StandardIcons
 	public const string FormatJustifyLeft = "format-justify-left-symbolic";
 	public const string FormatJustifyCenter = "format-justify-center-symbolic";
 	public const string FormatJustifyRight = "format-justify-right-symbolic";
-	public const string FormatTextBold = "format-text-bold-symbolic";
 	public const string FormatTextItalic = "format-text-italic-symbolic";
 	public const string FormatTextUnderline = "format-text-underline-symbolic";
 
@@ -185,6 +184,10 @@ public static class Icons
 	public const string OrientationPortrait = "image-orientation-portrait-symbolic";
 	public const string OrientationLandscape = "image-orientation-landscape-symbolic";
 
+	public const string JoinMiter = "join-miter-symbolic";
+	public const string JoinRound = "join-round-symbolic";
+	public const string JoinBevel = "join-bevel-symbolic";
+
 	public const string LayerDelete = "layers-remove-layer-symbolic";
 	public const string LayerDuplicate = "layers-duplicate-layer-symbolic";
 	public const string LayerFlipHorizontal = ImageFlipHorizontal;
@@ -214,6 +217,12 @@ public static class Icons
 	public const string Sampling5 = "tool-colorpicker-sampling-5x5-symbolic";
 	public const string Sampling7 = "tool-colorpicker-sampling-7x7-symbolic";
 	public const string Sampling9 = "tool-colorpicker-sampling-9x9-symbolic";
+
+	public const string TextExtraLight = "text-extra-light-symbolic";
+	public const string TextLight = "text-light-symbolic";
+	public const string TextNormal = "text-normal-symbolic";
+	public const string TextBold = "text-bold-symbolic";
+	public const string TextExtraBold = "text-extra-bold-symbolic";
 
 	public const string ToolCloneStamp = "tool-clonestamp-symbolic";
 	public const string ToolColorPicker = "tool-colorpicker-symbolic";