+++ Pinta.Tools/Tools/TextTool.cs
@@ -124,7 +124,8 @@ public sealed class TextTool : BaseTool
 	// NRT - Created by OnBuildToolBar
 	private Gtk.Label font_label = null!;
 	private Gtk.FontDialogButton font_button = null!;
-	private Gtk.ToggleButton bold_btn = null!;
+	private Gtk.SpinButton font_size = null!;
+	private ToolBarDropDownButton weight_btn = null!;
 	private Gtk.ToggleButton italic_btn = null!;
 	private Gtk.ToggleButton underscore_btn = null!;
 	private Gtk.ToggleButton left_alignment_btn = null!;
@@ -136,6 +137,8 @@ public sealed class TextTool : BaseTool
 	private Gtk.Separator outline_sep = null!;
 	private Gtk.SpinButton outline_width = null!;
 	private Gtk.Label outline_width_label = null!;
+	private Gtk.Separator join_sep = null!;
+	private ToolBarDropDownButton join_btn = null!;
 
 	protected override void OnBuildToolBar (Gtk.Box tb)
 	{
@@ -157,6 +160,7 @@ public sealed class TextTool : BaseTool
 				UseSize = false,
 				UseFont = true,
 				CanFocus = false,
+				Level = Gtk.FontLevel.Face,
 				FontDesc = Pango.FontDescription.FromString (
 					Settings.GetSetting (SettingNames.TEXT_FONT,
 					Gtk.Settings.GetDefault ()!.GtkFontName!)),
@@ -171,17 +175,106 @@ public sealed class TextTool : BaseTool
 
 		tb.Append (GtkExtensions.CreateToolBarSeparator ());
 
-		if (bold_btn == null) {
-			bold_btn = new Gtk.ToggleButton {
-				IconName = Pinta.Resources.StandardIcons.FormatTextBold,
-				TooltipText = Translations.GetString ("Bold"),
-				CanFocus = false,
-				Active = Settings.GetSetting (SettingNames.TEXT_BOLD, false),
+		if (font_size == null) {
+			var font_size_adjustment = new Gtk.Adjustment {
+				Lower = 1,
+				Upper = 2000,
+				StepIncrement = 1,
+				Value = PangoExtensions.UnitsToPixels (font_button.FontDesc!.GetSize ()),
 			};
-			bold_btn.OnToggled += HandleBoldButtonToggled;
+
+			font_size = new Gtk.SpinButton {
+				Adjustment = font_size_adjustment,
+				TooltipText = Translations.GetString ("Change font size. Shortcut keys: [ ]"),
+			};
+			font_size.OnValueChanged += HandleFontSizeChanged;
 		}
 
-		tb.Append (bold_btn);
+		tb.Append (font_size);
+
+		tb.Append (GtkExtensions.CreateToolBarSeparator ());
+
+		if (weight_btn == null) {
+			weight_btn = new ToolBarDropDownButton ();
+
+			// Translators: 'Thin' (100) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Thin") + " 100",
+				Pinta.Resources.Icons.TextExtraLight,
+				Pango.Weight.Thin
+			);
+			// Translators: 'Ultralight' (200) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Ultralight") + " 200",
+				Pinta.Resources.Icons.TextExtraLight,
+				Pango.Weight.Ultralight
+			);
+			// Translators: 'Light' (300) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Light") + " 300",
+				Pinta.Resources.Icons.TextLight,
+				Pango.Weight.Light
+			);
+			// Translators: 'Semilight' (350) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Semilight") + " 350",
+				Pinta.Resources.Icons.TextLight,
+				Pango.Weight.Semilight
+			);
+			// Translators: 'Book' (380) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Book") + " 380",
+				Pinta.Resources.Icons.TextNormal,
+				Pango.Weight.Book
+			);
+			// Translators: 'Normal' (400) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Normal") + " 400",
+				Pinta.Resources.Icons.TextNormal,
+				Pango.Weight.Normal
+			);
+			// Translators: 'Medium' (500) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Medium") + " 500",
+				Pinta.Resources.Icons.TextNormal,
+				Pango.Weight.Medium
+			);
+			// Translators: 'Semibold' (600) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Semibold") + " 600",
+				Pinta.Resources.Icons.TextBold,
+				Pango.Weight.Semibold
+			);
+			// Translators: 'Bold' (700) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Bold") + " 700",
+				Pinta.Resources.Icons.TextBold,
+				Pango.Weight.Bold
+			);
+			// Translators: 'Ultrabold' (800) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Ultrabold") + " 800",
+				Pinta.Resources.Icons.TextExtraBold,
+				Pango.Weight.Ultrabold
+			);
+			// Translators: 'Heavy' (900) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Heavy") + " 900",
+				Pinta.Resources.Icons.TextExtraBold,
+				Pango.Weight.Heavy
+			);
+			// Translators: 'Ultraheavy' (1000) refers to the font-weight text property
+			weight_btn.AddItem (
+				Translations.GetString ("Ultraheavy") + " 1000",
+				Pinta.Resources.Icons.TextExtraBold,
+				Pango.Weight.Ultraheavy
+			);
+
+			weight_btn.SelectedIndex = Settings.GetSetting (SettingNames.TEXT_WEIGHT, 5);
+			weight_btn.SelectedItemChanged += HandleWeightButtonToggled;
+		}
+
+		tb.Append (weight_btn);
 
 		if (italic_btn == null) {
 			italic_btn = new Gtk.ToggleButton {
@@ -267,7 +360,7 @@ public sealed class TextTool : BaseTool
 			fill_button.AddItem (Translations.GetString ("Fill Background"), Pinta.Resources.Icons.FillStyleBackground, 3);
 
 			fill_button.SelectedIndex = Settings.GetSetting (SettingNames.TEXT_STYLE, 0);
-			fill_button.SelectedItemChanged += HandleBoldButtonToggled;
+			fill_button.SelectedItemChanged += HandleFillButtonToggled;
 		}
 
 		tb.Append (fill_button);
@@ -294,7 +387,48 @@ public sealed class TextTool : BaseTool
 
 		tb.Append (outline_width);
 
-		outline_width.Visible = outline_width_label.Visible = outline_sep.Visible = StrokeText;
+		join_sep ??= GtkExtensions.CreateToolBarSeparator ();
+
+		tb.Append (join_sep);
+
+		if (join_btn == null) {
+			join_btn = new ToolBarDropDownButton ();
+
+			// Translators: 'Miter Join' refers to the Cairo.LineJoin property
+			join_btn.AddItem (
+				Translations.GetString ("Miter Join"),
+				Pinta.Resources.Icons.JoinMiter,
+				Cairo.LineJoin.Miter
+			);
+			// Translators: 'Round Join' refers to the Cairo.LineJoin property
+			join_btn.AddItem (
+				Translations.GetString ("Round Join"),
+				Pinta.Resources.Icons.JoinRound,
+				Cairo.LineJoin.Round
+			);
+			// Translators: 'Bevel Join' refers to the Cairo.LineJoin property
+			join_btn.AddItem (
+				Translations.GetString ("Bevel Join"),
+				Pinta.Resources.Icons.JoinBevel,
+				Cairo.LineJoin.Bevel
+			);
+
+			join_btn.SelectedIndex = Settings.GetSetting (SettingNames.TEXT_JOIN, 0);
+			join_btn.SelectedItemChanged += HandleJoinButtonToggled;
+		}
+
+		tb.Append (join_btn);
+
+		outline_width.Visible = outline_width_label.Visible = outline_sep.Visible = join_btn.Visible = join_sep.Visible = StrokeText;
+
+		UpdateFont ();
+	}
+
+	private void HandleFontSizeChanged (object? sender, EventArgs e)
+	{
+		var font = font_button.FontDesc!.Copy ()!;
+		font.SetSize (PangoExtensions.UnitsFromPixels (font_size.GetValueAsInt ()));
+		font_button.FontDesc = font;
 
 		UpdateFont ();
 	}
@@ -306,8 +440,8 @@ public sealed class TextTool : BaseTool
 		if (font_button is not null)
 			settings.PutSetting (SettingNames.TEXT_FONT, font_button.FontDesc!.ToString ()!);
 
-		if (bold_btn is not null)
-			settings.PutSetting (SettingNames.TEXT_BOLD, bold_btn.Active);
+		if (weight_btn is not null)
+			settings.PutSetting (SettingNames.TEXT_WEIGHT, weight_btn.SelectedIndex);
 
 		if (italic_btn is not null)
 			settings.PutSetting (SettingNames.TEXT_ITALIC, italic_btn.Active);
@@ -323,10 +457,17 @@ public sealed class TextTool : BaseTool
 
 		if (outline_width is not null)
 			settings.PutSetting (SettingNames.TEXT_OUTLINE_WIDTH, outline_width.GetValueAsInt ());
+
+		if (join_btn is not null)
+			settings.PutSetting (SettingNames.TEXT_JOIN, join_btn.SelectedIndex);
 	}
 
 	private void HandleFontChanged ()
 	{
+		var font = font_button.FontDesc!.Copy ()!;
+		font.SetSize (PangoExtensions.UnitsFromPixels (font_size.GetValueAsInt ()));
+		font_button.FontDesc = font;
+
 		if (workspace.HasOpenDocuments)
 			workspace.ActiveDocument.Workspace.GrabFocusToCanvas ();
 
@@ -397,10 +538,20 @@ public sealed class TextTool : BaseTool
 		UpdateFont ();
 	}
 
-	private void HandleBoldButtonToggled (object? sender, EventArgs e)
+	private void HandleWeightButtonToggled (object? sender, EventArgs e)
 	{
-		outline_width.Visible = outline_width_label.Visible = outline_sep.Visible = StrokeText;
+		UpdateFont ();
+	}
 
+	private void HandleFillButtonToggled (object? sender, EventArgs e)
+	{
+		outline_width.Visible = outline_width_label.Visible = outline_sep.Visible = join_btn.Visible = join_sep.Visible = StrokeText;
+
+		UpdateFont ();
+	}
+
+	private void HandleJoinButtonToggled (object? sender, EventArgs e)
+	{
 		UpdateFont ();
 	}
 
@@ -419,7 +570,7 @@ public sealed class TextTool : BaseTool
 		if (workspace.HasOpenDocuments) {
 
 			var font = font_button.FontDesc!.Copy ()!; // NRT: Only nullable when nullptr is passed.
-			font.SetWeight (bold_btn.Active ? Pango.Weight.Bold : Pango.Weight.Normal);
+			font.SetWeight ((Pango.Weight) weight_btn.SelectedItem.GetTagOrDefault (Pango.Weight.Normal));
 			font.SetStyle (italic_btn.Active ? Pango.Style.Italic : Pango.Style.Normal);
 
 			CurrentTextEngine.SetFont (font, Alignment, underscore_btn.Active);
@@ -762,7 +913,8 @@ public sealed class TextTool : BaseTool
 								italic_btn.Toggle ();
 								UpdateFont ();
 							} else if (e.Key.Value == Gdk.Constants.KEY_b) {
-								bold_btn.Toggle ();
+								// If current font-weight is Bold (8) or bolder, set to Normal (5). Otherwise, set to Bold (8).
+								weight_btn.SelectedIndex = weight_btn.SelectedIndex > 7 ? 5 : 8;
 								UpdateFont ();
 							} else if (e.Key.Value == Gdk.Constants.KEY_u) {
 								underscore_btn.Toggle ();
@@ -786,6 +938,15 @@ public sealed class TextTool : BaseTool
 
 			if (keyHandled)
 				RedrawText (true, true);
+		} else {
+			switch (e.Key.Value) {
+				case Gdk.Constants.KEY_bracketleft:
+					font_size.Adjustment!.Value--;
+					return true;
+				case Gdk.Constants.KEY_bracketright:
+					font_size.Adjustment!.Value++;
+					return true;
+			}
 		}
 
 		return keyHandled;
@@ -1032,6 +1193,7 @@ public sealed class TextTool : BaseTool
 		if (StrokeText) {
 			g.SetSourceColor (FillText ? CurrentTextEngine.SecondaryColor : CurrentTextEngine.PrimaryColor);
 			g.LineWidth = OutlineWidth;
+			g.LineJoin = (Cairo.LineJoin) join_btn.SelectedIndex;
 
 			PangoCairo.Functions.LayoutPath (g, CurrentTextLayout.Layout);
 			g.Stroke ();