+++ Pinta/ActionHandlers.cs
@@ -1,21 +1,21 @@
-// 
+//
 // FileActionHandler.cs
-// 
+//
 // Author:
 //       Jonathan Pobst <monkey@jpobst.com>
-// 
+//
 // Copyright (c) 2010 Jonathan Pobst
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -70,8 +70,8 @@ public sealed class ActionHandlers
 			new AddinManagerAction (actions.Addins, chrome, system),
 
 			// Image
-			new ResizeImageAction (actions.Image, chrome, workspace),
-			new ResizeCanvasAction (chrome, workspace, actions),
+			new ResizeImageAction (actions.Image, chrome, workspace, settings),
+			new ResizeCanvasAction (chrome, workspace, settings, actions),
 
 			// Layers
 			new LayerPropertiesAction (chrome, actions.Layers, workspace),